% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{add_badge_comments}
\alias{add_badge_comments}
\title{Add special marker comments delimiting the badge section to package README}
\usage{
add_badge_comments(x = "README.md", f = NULL)
}
\arguments{
\item{x}{Name of README file}

\item{f}{Absolute path to file (can be used instead of \code{x} for
debugging)}
}
\description{
Add special marker comments delimiting the badge section to package README
}
\details{
Will only add badge comments if they are not yet present. If there
  are some badges present then will add marker comments on either side. Will
  give a warning and do noohting if badge lines are not contiguous
}
