% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat.devtools-package.R
\docType{package}
\name{nat.devtools-package}
\alias{nat.devtools}
\alias{nat.devtools-package}
\title{eywords internal}
\description{
*nat.devtools* is built on top of the \code{\link{usethis}} package. It
  shares the key concept of the active package. If you are using Rstudio to
  develop a package, then all should be fine. You can start by running
  \code{\link{nat_setup_package}}, which should do everything for a complete
  package setup.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/nat.devtools}
  \item Report bugs at \url{https://github.com/jefferis/nat.devtools/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
