% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo_utils.R
\name{nat_hugo_addarticle}
\alias{nat_hugo_addarticle}
\title{Create an article from a rmarkdown file and add it in a hugo repository}
\usage{
nat_hugo_addarticle(urlpath, foldername, title, author, categories,
  description, slug, logourl_path, metadata_photo)
}
\arguments{
\item{urlpath}{Path to the rmarkdown file (usually a location on the web)}

\item{foldername}{name of the folder to be created under `/content/articles/` like '2019-09-nat'}

\item{title}{Set the title of the article like 'article_xxx'}

\item{author}{Set the author name like sri, greg etc (corresponding to names in
the hugo repository)}

\item{categories}{Set the category of the article like 'learn', 'nat' etc.}

\item{description}{Set the description of the article like '`nat` version 1.8.13 is now on CRAN!'}

\item{slug}{Set the slug of the article like 'nat-1-8-13'}

\item{logourl_path}{Path to the square logo file that is displayed as favicon (usually a location on the web)}

\item{metadata_photo}{Set the metadata related to the photo to appear on the article (must be a wide figure)
like c(photoauthor, photo_url)}
}
\description{
Create an article from a rmarkdown file and add it in a hugo repository
}
\examples{
\dontrun{
nat_hugo_addarticle(urlpath='https://raw.githubusercontent.com/natverse/nat/master/vignettes/NeuroGeometry.Rmd',
foldername = '2019-09-nat', title = 'NeuroGeometry - Analysing 3D morphology of neurons',
author = c('greg'), categories = c('learn','nat'), description = '`nat` version 1.8.13 is now on CRAN!',
slug = 'nat-1-8-13', logourl_path = 'https://raw.githubusercontent.com/natverse/natverse_hugo/dev_pages/content/articles/2019-09-nat/natverse_logo-sq.jpg',
metadata_photo = c('Maranda Vandergriff', 'https://unsplash.com/photos/o-d37kiKqqc'))

nat_hugo_addarticle(urlpath='https://raw.githubusercontent.com/natverse/nat/master/vignettes/neurons-intro.Rmd',
foldername = '2019-09-neurons', title = 'Introduction to neurons and neuronlists',
author = c('greg'), categories = c('learn','nat'), description = '`nat` neurons and neuronlists!',
slug = 'nat',logourl_path = 'https://raw.githubusercontent.com/natverse/natverse_hugo/dev_pages/content/articles/2019-09-nat/natverse_logo-sq.jpg',
metadata_photo = c('Wikipedia', 'https://upload.wikimedia.org/wikipedia/commons/4/45/Wide_lightning.jpg'))
}
}
