% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hexlogo.R
\name{nat_makehexlogo}
\alias{nat_makehexlogo}
\title{Make hexagon logo from any kind of image}
\usage{
nat_makehexlogo(rawimgpath, outputimgpath, hexbordercolour = "white",
  hexinnercolour = "red")
}
\arguments{
\item{rawimgpath}{Path to the png file (that contains the actual logo)}

\item{outputimgpath}{Path to the output png file that is in the form of a hex}

\item{hexbordercolour}{Colour of the hexagon border}

\item{hexinnercolour}{Colour to fill inside (usually give the background color of your image)}
}
\description{
This code was inspired and adapted from the make hexagon code in https://github.com/greta-dev/greta/blob/master/logos/make_hex.R
}
\examples{
\dontrun{
nat_makehexlogo(rawimgpath='logos/nat_withouthex.png',outputimgpath ='logos/nat_withhex.png')
}
}
