% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{nat_setup_badges}
\alias{nat_setup_badges}
\title{Adds all standard badges to your README.md}
\usage{
nat_setup_badges(x = "README.md", lifecycle = NULL)
}
\arguments{
\item{x}{Name of README file}

\item{lifecycle}{The developmental stage of the package (see
\url{https://www.tidyverse.org/lifecycle})}
}
\description{
Adds all standard badges to your README.md
}
\seealso{
\code{\link{natverse-badges}}
}
