% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-utils.R
\name{nv_check_urls}
\alias{nv_check_urls}
\title{Check for files containing old jefferis/lab URLs.}
\usage{
nv_check_urls(path = ".",
  pattern = c("^(DESCRIPTION|NEWS|.*\\\\.(R|r|Rmd|md|yml))$"),
  oldusers = c("jefferis", "jefferislab", "flyconnectome"),
  replace = FALSE)
}
\arguments{
\item{path}{Path to git repository containing package (defaults to current
working directory)}

\item{pattern}{Files to check (passed to \code{\link{dir}})}

\item{oldusers}{Which GitHub users need changing to natverse}

\item{replace}{Whether to replace old URLs with natverse URLs}
}
\value{
Paths to any bad files
}
\description{
Check for files containing old jefferis/lab URLs.
}
