% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{use_nat_description_defaults}
\alias{use_nat_description_defaults}
\title{Set default authorship/package DESCRIPTION information in your Rprofile}
\usage{
use_nat_description_defaults()
}
\description{
Set default authorship/package DESCRIPTION information in your Rprofile
}
\seealso{
See \code{usethis::\link[usethis]{use_description_defaults}} and
  \code{\link{Rprofile}}
}
