% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-utils.R
\name{use_nat_support}
\alias{use_nat_support}
\title{Add a SUPPORT.md file with information about getting help}
\usage{
use_nat_support()
}
\description{
Add a SUPPORT.md file with information about getting help
}
