% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{add_badge_comments}
\alias{add_badge_comments}
\title{Add special marker comments delimiting the badge section of package README}
\usage{
add_badge_comments(x = "README.md", f = NULL)
}
\arguments{
\item{x}{Name of README file}

\item{f}{Absolute path to file (can be used instead of \code{x} for
debugging)}
}
\value{
Path to modified readme file (invisibly)
}
\description{
Add special marker comments delimiting the badge section of package README
}
\details{
Badge comment lines are required for
  \code{usethis::\link{use_badge}} and friends to automagically add badges to
  a README.

  \code{add_badge_comments} will only add badge comments if they are not yet
  present. If there are some badges present then will add marker comments on
  either side. Will give a warning and do nothing if badge lines are not
  contiguous
}
\seealso{
\code{\link{natverse-badges}}, \code{usethis::\link{use_badge}}
}
