% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat.devtools-package.R
\docType{package}
\name{nat.devtools-package}
\alias{nat.devtools}
\alias{nat.devtools-package}
\title{nat.devtools: Tools to help develop natverse packages}
\description{
Provides functions to automate some aspects of
    package setup for the natverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/nat.devtools}
  \item Report bugs at \url{https://github.com/jefferis/nat.devtools/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
\keyword{internal}
