% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat.devtools-package.R
\docType{package}
\name{nat.devtools-package}
\alias{nat.devtools}
\alias{nat.devtools-package}
\title{nat.devtools: Tools to help develop natverse packages}
\description{
Provides functions to automate some aspects of
    package setup for the natverse.
}
\details{
\bold{nat.devtools} is built on top of the \code{\link{usethis}}
  package. It shares the key concept of the active package. If you are using
  Rstudio to develop a single package, then all should be fine. If you need
  to change between different packages, check out the
  \code{usethis::\link{proj_utils}} dcoumentation.

  You can start by running \code{\link{nat_setup_package}}, which should do
  everything for a complete package setup. More specialised functions are
  available for selected tasks once you already have a package setup.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://natverse.org/nat.devtools}
  \item \url{https://github.com/natverse/nat.devtools}
  \item Report bugs at \url{https://github.com/natverse/nat.devtools/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
