% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-utils.R
\name{nat_setup_package}
\alias{nat_setup_package}
\title{Setup a package in natverse style}
\usage{
nat_setup_package(
  path = ".",
  lifecycle = c("experimental", "maturing", "stable")
)
}
\arguments{
\item{path}{Path to git repository containing package (defaults to current
working directory)}

\item{lifecycle}{The developmental stage of the package (see
\url{https://www.tidyverse.org/lifecycle})}
}
\description{
Setup a package in natverse style
}
\examples{
\dontrun{
nat.devtools::nat_setup_package(lifecycle='maturing')
nat.devtools::nat_setup_package(lifecycle='stable')
nat.devtools::nat_setup_package(lifecycle='experimental')
}
}
\seealso{
\code{\link{nat_setup_badges}}, \code{\link{nat_setup_pkgdown}}
}
