% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{nat_setup_pkgdown}
\alias{nat_setup_pkgdown}
\title{Setup pkgdown according to natverse conventions}
\usage{
nat_setup_pkgdown(config_file = "_pkgdown.yml", github = TRUE)
}
\arguments{
\item{config_file}{The name of the pkgdown config file (default
\code{"_pkgdown.yml"})}

\item{github}{Whether to use github to build the pkgdown site (default
\code{TRUE})}
}
\value{
a \code{list} describing the pkgdown config
}
\description{
Setup pkgdown according to natverse conventions
}
