% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\name{natverse-badges}
\alias{natverse-badges}
\alias{use_doc_badge}
\alias{use_cran_downloads_badge}
\alias{use_natverse_badge}
\alias{nat_standard_badges}
\title{Add natverse badges to package README}
\usage{
use_doc_badge()

use_cran_downloads_badge()

use_natverse_badge()

nat_standard_badges(x = "README.md", lifecycle = NULL)
}
\arguments{
\item{x}{Name of README file}

\item{lifecycle}{The developmental stage of the package (see
\url{https://www.tidyverse.org/lifecycle})}
}
\description{
\code{use_doc_badge} adds a badge linking to the pkgdown site
  for the current package

\code{use_cran_downloads_badge} adds a badge to give the number
  of downloads / month from CRAN.

\code{use_natverse_badge} adds a badge to say that package is a member of the natverse

\code{nat_standard_badges} Adds all standard badges to your README.md
}
